Move RuntimeStats into separate source files
